+++ spaceinvaderspwng/code/space_invaders/src/globals.h
@@ -49,8 +49,8 @@ void globals_setAlienBlockPosition(point_t val);
 point_t globals_getAlienBlockPosition();
 
 // set and get the position of the saucer
-void globals_setSaucerPosition(uint16_t val);
-uint32_t globals_getSaucerPosition();
+void globals_setSaucerPosition(int32_t val);
+int32_t globals_getSaucerPosition();
 
 // set and get the position of the specified alien bullet
 void globals_setAlienBulletPosition(point_t val, uint8_t bullet);
